+++ src/main/java/classFile/constants/ConstantDoubleInfo.java
@@ -6,7 +6,7 @@ public class ConstantDoubleInfo extends ConstantInfo {
 
     private double value;
 
-    public ConstantDoubleInfo(byte tag, ByteReader reader) {
+    public ConstantDoubleInfo(int tag, ByteReader reader) {
         super(tag);
         long longBits = reader.readUint64();
         this.value = Double.longBitsToDouble(longBits);
