+++ src/main/java/classFile/constants/ConstantFloatInfo.java
@@ -6,7 +6,7 @@ public class ConstantFloatInfo extends ConstantInfo {
 
     private float value;
 
-    public ConstantFloatInfo(byte tag, ByteReader reader) {
+    public ConstantFloatInfo(int tag, ByteReader reader) {
         super(tag);
         int intBits = reader.readUint32();
         this.value = Float.intBitsToFloat(intBits);
