+++ src/main/java/classFile/constants/ConstantIntegerInfo.java
@@ -5,7 +5,7 @@ import classFile.reader.ByteReader;
 public class ConstantIntegerInfo extends ConstantInfo {
     private int value;
 
-    public ConstantIntegerInfo(byte tag, ByteReader reader) {
+    public ConstantIntegerInfo(int tag, ByteReader reader) {
         super(tag);
         this.value = reader.readUint32();
     }
