+++ src/main/java/classFile/constants/ConstantInterfaceMethodrefInfo.java
@@ -4,20 +4,20 @@ import classFile.reader.ByteReader;
 
 public class ConstantInterfaceMethodrefInfo extends ConstantInfo {
 
-    private short classIndex;
-    private short nameAndTypeIndex;
+    private int classIndex;
+    private int nameAndTypeIndex;
 
-    public ConstantInterfaceMethodrefInfo(byte tag, ByteReader reader) {
+    public ConstantInterfaceMethodrefInfo(int tag, ByteReader reader) {
         super(tag);
         this.classIndex = reader.readUnit16();
         this.nameAndTypeIndex = reader.readUnit16();
     }
 
-    public short getClassIndex() {
+    public int getClassIndex() {
         return classIndex;
     }
 
-    public short getNameAndTypeIndex() {
+    public int getNameAndTypeIndex() {
         return nameAndTypeIndex;
     }
 }
