+++ src/main/java/classFile/constants/ConstantLongInfo.java
@@ -5,7 +5,7 @@ import classFile.reader.ByteReader;
 public class ConstantLongInfo extends ConstantInfo {
     private long value;
 
-    public ConstantLongInfo(byte tag, ByteReader reader) {
+    public ConstantLongInfo(int tag, ByteReader reader) {
         super(tag);
         this.value = reader.readUint64();
     }
