+++ src/main/java/classFile/constants/ConstantMethodHandleInfo.java
@@ -4,20 +4,20 @@ import classFile.reader.ByteReader;
 
 public class ConstantMethodHandleInfo extends ConstantInfo {
 
-    private byte referenceKind;
-    private short referenceIndex;
+    private int referenceKind;
+    private int referenceIndex;
 
-    public ConstantMethodHandleInfo(byte tag, ByteReader reader) {
+    public ConstantMethodHandleInfo(int tag, ByteReader reader) {
         super(tag);
         this.referenceKind = reader.readUnit8();
         this.referenceIndex = reader.readUnit16();
     }
 
-    public byte getReferenceKind() {
+    public int getReferenceKind() {
         return referenceKind;
     }
 
-    public short getReferenceIndex() {
+    public int getReferenceIndex() {
         return referenceIndex;
     }
 }
