+++ src/main/java/classFile/constants/ConstantMethodrefInfo.java
@@ -4,20 +4,20 @@ import classFile.reader.ByteReader;
 
 public class ConstantMethodrefInfo extends ConstantInfo {
 
-    private short classIndex;
-    private short nameAndTypeIndex;
+    private int classIndex;
+    private int nameAndTypeIndex;
 
-    public ConstantMethodrefInfo(byte tag, ByteReader reader) {
+    public ConstantMethodrefInfo(int tag, ByteReader reader) {
         super(tag);
         this.classIndex = reader.readUnit16();
         this.nameAndTypeIndex = reader.readUnit16();
     }
 
-    public short getClassIndex() {
+    public int getClassIndex() {
         return classIndex;
     }
 
-    public short getNameAndTypeIndex() {
+    public int getNameAndTypeIndex() {
         return nameAndTypeIndex;
     }
 }
