+++ src/main/java/classFile/constants/ConstantNameAndTypeInfo.java
@@ -4,21 +4,21 @@ import classFile.reader.ByteReader;
 
 public class ConstantNameAndTypeInfo extends ConstantInfo {
     // 字段名或者方法名在常量池索引
-    private short nameIndex;
+    private int nameIndex;
     // 字段名或者方法的描述符常量池索引
-    private short descriptorIndex;
+    private int descriptorIndex;
 
-    public ConstantNameAndTypeInfo(byte tag, ByteReader reader) {
+    public ConstantNameAndTypeInfo(int tag, ByteReader reader) {
         super(tag);
         this.nameIndex = reader.readUnit16();
         this.descriptorIndex = reader.readUnit16();
     }
 
-    public short getNameIndex() {
+    public int getNameIndex() {
         return nameIndex;
     }
 
-    public short getDescriptorIndex() {
+    public int getDescriptorIndex() {
         return descriptorIndex;
     }
 }
