+++ src/main/java/classFile/constants/ConstantUtf8Info.java
@@ -9,9 +9,9 @@ public class ConstantUtf8Info extends ConstantInfo {
 
     private String value;
 
-    public ConstantUtf8Info(byte tag, ByteReader reader) {
+    public ConstantUtf8Info(int tag, ByteReader reader) {
         super(tag);
-        short len = reader.readUnit16();
+        int len = reader.readUnit16();
         byte[] bytes = reader.readBytes(len);
         try {
             this.value = MUTF8.toUtf8String(bytes);
